Renaming the project to WeinerBot. I don't want the project to have a political affiliation anymore. All of the political functionality has been removed in previous commits as it's no longer useful and I'm completely burned out on politics now.
WeinerBot is a dumb name but it's also a dumb bot so what do you expect?
It has been a long 5 1/2 years, I originally created the project in
January 2016 and it's by far my longest running project and I still
intend on making improvements and creating new functionality.

RE: CuckHunt, this will be reworked eventually.
I'm looking at either redesigning it or removing the module altogether
and exposing the stats functionality as a command. Essentially leaving
it in a read-only state.